stopping about halfway of step 4
Decided to spend final full day to prepare for live project with other resources. Spent about 2 days working on this and I was running into issues that are caused by updates since 2017, which is when this tutorial was from. Was very helpful and gave me some more experience working in it.